initial sketch of main tests

+++ src/automat/utils/InputStream.java
@@ -4,5 +4,5 @@ import java.io.IOException;
 
 public interface InputStream {
     Object nextInput(Object eof) throws IOException;
-    int nextNumericInput(int eof) throws IOException;
+    long nextNumericInput(long eof) throws IOException;
 }
